# API/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "DeepLearning"
  },
  "paths": {
    "/signup": {
      "post": {
        "summary": "Inscription",
        "description": "Créer un compte",
        "parameters": [
          {
            "name": "username",
            "in": "header",
            "description": "Username",
            "required": true,
            "type": "string",
            "format": "string"
          },
          {
            "name": "password",
            "in": "header",
            "description": "Password",
            "required": true,
            "type": "string",
            "format": "password"
          }
        ],
        "responses": {
          "200": {
            "description": "Opération réussie"
          },
          "409": {
            "description": "L'utilisateur existe déjà"
          },
          "400": {
            "description": "Mauvaise requête"
          }
        }
      }
    },
    "/signin": {
      "get": {
        "summary": "Connexion",
        "description": "Pour s'authentifier",
        "parameters": [
          {
            "name": "username",
            "in": "header",
            "description": "Username",
            "required": true,
            "type": "string",
            "format": "string"
          },
          {
            "name": "password",
            "in": "header",
            "description": "Password",
            "required": true,
            "type": "string",
            "format": "password"
          }
        ],
        "responses": {
          "200": {
            "description": "Opération réussie"
          },
          "404": {
            "description": "Username ou mot de passe incorrect"
          }
        }
      }
    },"/checkToken": {
      "get": {
        "summary": "Vérifie le token",
        "description": "Vérifie la validité d'un token",
        "parameters": [
          {
            "name": "sentence",
            "in": "query",
            "description": "Token",
            "required": true,
            "type": "string",
            "format": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Token valide"
          },
          "401": {
            "description": "Token invalide"
          }
        }
      }
    },
    "/getInfo": {
      "get": {
        "summary": "Informations sur l'utilisateur",
        "description": "Informations sur l'utilisateur",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "description": "Token",
            "required": true,
            "type": "string",
            "format": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Opération réussie"
          },
          "403": {
            "description": "Token non valide"
          }
        }
      }
    },
    "/addWord": {
      "put": {
        "summary": "Ajouter un mot",
        "description": "Ajouter un mot",
        "parameters": [
          {
            "name": "word",
            "in": "query",
            "description": "Word",
            "required": true,
            "type": "string",
            "format": "string"
          },
          {
            "name": "token",
            "in": "header",
            "description": "Token",
            "required": true,
            "type": "string",
            "format": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Opération réussie"
          },
          "400": {
            "description": "Mauvaise requête"
          },
          "403": {
            "description": "Mauvais token"
          },
          "409": {
            "description": "Mot déjà dans la liste"
          }
        }
      }
    },
    "/delWord": {
      "put": {
        "summary": "Supprimer un mot",
        "description": "Supprimer un mot",
        "parameters": [
          {
            "name": "word",
            "in": "query",
            "description": "Word",
            "required": true,
            "type": "string",
            "format": "string"
          },
          {
            "name": "token",
            "in": "header",
            "description": "Token",
            "required": true,
            "type": "string",
            "format": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Opération réussie"
          },
          "400": {
            "description": "Mauvaise requête"
          },
          "403": {
            "description": "Mauvais token"
          },
          "409": {
            "description": "Aucun mot supprimé"
          }
        }
      }
    },
    "/modifyWord": {
      "put": {
        "summary": "Modifier un mot",
        "description": "Modifier un mot",
        "parameters": [
          {
            "name": "word",
            "in": "query",
            "description": "Word",
            "required": true,
            "type": "string",
            "format": "string"
          },
          {
            "name": "token",
            "in": "header",
            "description": "Token",
            "required": true,
            "type": "string",
            "format": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Opération réussie"
          },
          "400": {
            "description": "Mauvaise requête"
          },
          "403": {
            "description": "Mauvais token"
          },
          "409": {
            "description": "Aucun mot modifié"
          }
        }
      }
    },
    "/emptyList": {
      "put": {
        "summary": "Supprimer la liste de vocabulaire",
        "description": "Supprimer la liste de vocabulaire",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "description": "Token",
            "required": true,
            "type": "string",
            "format": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Opération réussie"
          },
          "403": {
            "description": "Opération échouée"
          }
        }
      }
    },
    "/update": {
      "put": {
        "summary": "Met à jour la liste de vocabulaire",
        "description": "Met à jour la liste de vocabulaire pour chaque mot non traduit",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "description": "Token",
            "required": true,
            "type": "string",
            "format": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Opération réussie"
          }
        }
      }
    }
  }
}